@@ -26,35 +26,36 @@
     "shapefile"
   ],
   "author": "Christos Charmatzis",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/Charmatzis/react-leaflet-shapefile/issues"
   },
   "homepage": "https://github.com/Charmatzis/react-leaflet-shapefile#readme",
   "peerDependencies": {
     "catiline": "2.9.3",
-    "jquery": "*",
-    "leaflet": "^0.7.7",
-    "react-leaflet": "^0.11.5",
-    "react": "^15.0.2"
+    "leaflet": "^1.0.2",
+    "react-leaflet": "^1.0.1",
+    "react": "^15.4.1",
+    "shpjs": "^3.3.2"
   },
   "devDependencies": {
     "babel-cli": "^6.8.0",
     "babel-core": "^6.8.0",
-    "babel-eslint": "^4.1.8",
+    "babel-eslint": "^7.0.0",
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.0.15",
     "babel-preset-stage-0": "^6.1.18",
     "catiline": "2.9.3",
-    "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.16.1",
+    "eslint": "^3.8.1",
+    "eslint-plugin-react": "^6.4.1",
     "jquery": "*",
-    "leaflet": "^0.7.7",
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2",
-    "react-file-drop": "*",
-    "react-leaflet": "^0.11.5",
+    "leaflet": "^1.0.2",
+    "react": "^15.4.1",
+    "requirejs": "^2.3.2",
+    "react-dom": "^15.4.1",
+    "react-leaflet": "^1.0.1",    
+    "shpjs": "^3.3.2",
     "webpack": "^1.13.0",
     "webpack-dev-server": "^1.12.1"
   }
